+++ HISTORY.md
@@ -3,6 +3,7 @@
     * Added support for system dbus files
     * Added support for configuration dbus files
     * Fixed bug with aliases
+    * Fixed bug in meson when installing data files
 * version 1.1.2 (2017-10-16)
     * Improved the support for Meson build system (thanks to cromerc for the invaluable help)
     * Fixed the parallel compilation under CMake