+++ HISTORY.md
@@ -4,6 +4,7 @@
     * Added support for configuration dbus files
     * Fixed bug with aliases
     * Fixed bug in meson when installing data files
+    * Now doesn't add conditionals to the *translate* statement when a source file is conditional
 * version 1.1.2 (2017-10-16)
     * Improved the support for Meson build system (thanks to cromerc for the invaluable help)
     * Fixed the parallel compilation under CMake