+++ HISTORY.md
@@ -1,7 +1,10 @@
 ## History of versions ##
 * version 1.1.3 (2017-10-29)
     * Added support for system dbus files
+    * Added support for configuration dbus files
     * Fixed bug with aliases
+    * Fixed bug in meson when installing data files
+    * Now doesn't add conditionals to the *translate* statement when a source file is conditional
 * version 1.1.2 (2017-10-16)
     * Improved the support for Meson build system (thanks to cromerc for the invaluable help)
     * Fixed the parallel compilation under CMake